[file-interface][m]: Refactor file-interface to new script
